# tdi-service-tdis/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Servicio de Gestión de Actividades e Integración de TDIs.",
        "title": "TDI Service API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/api/tdi/alumnos/progreso": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Retorna la meta de puntos, puntos totales y el desglose de puntos por cada dimensión del alumno logueado.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "alumnos-info"
                ],
                "summary": "Obtener progreso de puntos del alumno",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID del usuario (Alumno)",
                        "name": "X-User-Id",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Progreso de puntos",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "No autorizado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/tdi/catalogo": {
            "get": {
                "description": "Obtiene todas las actividades vigentes con filtros opcionales",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tdi-catalog"
                ],
                "summary": "Listar y filtrar catálogo de TDIs",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filtrar por ID de Categoría",
                        "name": "categoria_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filtrar por ID de Dimensión",
                        "name": "dimension_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filtrar por ID de Entorno",
                        "name": "entorno_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filtrar por ID de Trascendencia",
                        "name": "trascendencia_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Buscador por nombre o descripción",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/db.CatalogItem"
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Inserta una actividad al catálogo general verificando el rol del usuario en la base de datos",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tdi-catalog"
                ],
                "summary": "Registrar nueva TDI en catálogo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID del usuario que ejecuta la acción",
                        "name": "X-User-Id",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Datos de la actividad",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.saveTDIReq"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Actividad creada",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "No autorizado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/tdi/catalogo/{id}": {
            "get": {
                "description": "Obtiene los detalles de una actividad específica del catálogo",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tdi-catalog"
                ],
                "summary": "Obtener detalle de una TDI del catálogo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID de la TDI",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/db.CatalogItem"
                        }
                    },
                    "404": {
                        "description": "No encontrada",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "Actualiza los campos de una actividad específica del catálogo verificando el rol del usuario en la BD",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tdi-catalog"
                ],
                "summary": "Editar una TDI del catálogo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID del usuario que ejecuta la acción",
                        "name": "X-User-Id",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID de la TDI a editar",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Datos actualizados",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.saveTDIReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Actividad actualizada",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "Elimina físicamente o expira lógicamente una actividad si ya tiene alumnos registrados",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tdi-catalog"
                ],
                "summary": "Eliminar una TDI del catálogo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Rol de usuario",
                        "name": "X-User-Role",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID de la TDI a borrar",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Actividad eliminada",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "db.CatalogItem": {
            "type": "object",
            "properties": {
                "categoria_id": {
                    "type": "string"
                },
                "creador_id": {
                    "type": "string"
                },
                "descripcion": {
                    "type": "string"
                },
                "dimension_id": {
                    "type": "string"
                },
                "entorno_id": {
                    "type": "string"
                },
                "evidencia_requerida": {
                    "type": "string"
                },
                "fecha_creacion": {
                    "type": "string"
                },
                "fecha_vencimiento": {
                    "type": "string"
                },
                "horas": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "nombre": {
                    "type": "string"
                },
                "puntaje": {
                    "type": "integer"
                },
                "trascendencia_id": {
                    "type": "string"
                }
            }
        },
...
        "handlers.saveTDIReq": {
            "type": "object",
            "properties": {
                "categoria_id": {
                    "type": "string"
                },
                "descripcion": {
                    "type": "string"
                },
                "dimension_id": {
                    "type": "string"
                },
                "entorno_id": {
                    "type": "string"
                },
                "evidencia_requerida": {
                    "type": "string"
                },
                "fecha_vencimiento": {
                    "description": "Formato AAAA-MM-DD",
                    "type": "string"
                },
                "horas": {
                    "type": "integer"
                },
                "nombre": {
                    "type": "string"
                },
                "puntaje": {
                    "type": "integer"
                },
                "trascendencia_id": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Escribe 'Bearer ' seguido de tu token JWT para autenticarte.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}